refactor: unify ResourceConfig and resolve_resources into shared module
ResourceConfig was defined identically in both agents.py and tools.py,
and _resolve_resources/_resolve_tool_resources were near-duplicate
merge-over-defaults helpers. Move both into app/models/shipwright.py
(which already holds other cross-router models like
ResourceConfigFromBuild) as ResourceConfig and resolve_resources(),
and have both routers import them.

Addresses review feedback on PR #2370.
